Move BalancedSegmentationDataset
to avoid circular import
