+++ src/unicorn_eval/adaptors/segmentation/data_handling.py
@@ -14,6 +14,7 @@
 from __future__ import annotations
 
 import logging
+import random
 from typing import Iterable
 
 import numpy as np
@@ -24,8 +25,6 @@ from torch.utils.data import Dataset
 from torch.utils.data._utils.collate import default_collate
 
 from unicorn_eval.adaptors.patch_extraction import extract_patches
-from unicorn_eval.adaptors.segmentation.baseline_segmentation_upsampling_3d.v2.data_handling import \
-    BalancedSegmentationDataset
 
 
 def assign_mask_to_patch(mask_data, x_patch, y_patch, patch_size, padding_value=0):
@@ -304,3 +303,67 @@ def load_patch_data(data_array: np.ndarray, batch_size: int = 80, balance_bg: bo
         train_ds = dataset_monai(data=data_array)
 
     return dataloader_monai(train_ds, batch_size=batch_size, shuffle=False)
+
+
+class BalancedSegmentationDataset(Dataset):
+    """
+    Balanced dataset for segmentation that ensures equal probability of sampling
+    positive and negative patches using inverse probability weighting.
+    """
+
+    def __init__(self, data, transform=None, random_seed=42):
+        self.transform = transform
+        self.rng = random.Random(random_seed)
+    
+        # Separate positive and negative patches
+        self.positive_patches = []
+        self.negative_patches = []
+    
+        for data_dict in data:
+            patch_label = data_dict["patch_label"]
+            if np.any(patch_label != 0):
+                self.positive_patches.append(data_dict)
+            else:
+                self.negative_patches.append(data_dict)
+
+        self.num_positive = len(self.positive_patches)
+        self.num_negative = len(self.negative_patches)
+
+        # Total length is twice the minimum class size to ensure balance
+        self.total_length = 2 * min(self.num_positive, self.num_negative) if min(self.num_positive, self.num_negative) > 0 else max(self.num_positive, self.num_negative)
+
+        print(f"BalancedSegmentationDataset: {self.num_positive} positive, {self.num_negative} negative patches")
+        print(f"Total balanced dataset size: {self.total_length}")
+
+    def __len__(self):
+        return self.total_length
+
+    def __getitem__(self, idx):
+        # Use inverse probability weighting: sample positive and negative with equal probability
+        if self.num_positive > 0 and self.num_negative > 0:
+            # 50% chance of sampling positive or negative
+            if self.rng.random() < 0.5:
+                # Sample positive patch
+                patch_idx = self.rng.randint(0, self.num_positive - 1)
+                data_dict = self.positive_patches[patch_idx]
+            else:
+                # Sample negative patch
+                patch_idx = self.rng.randint(0, self.num_negative - 1)
+                data_dict = self.negative_patches[patch_idx]
+        elif self.num_positive > 0:
+            # Only positive patches available
+            patch_idx = self.rng.randint(0, self.num_positive - 1)
+            data_dict = self.positive_patches[patch_idx]
+        else:
+            # Only negative patches available
+            patch_idx = self.rng.randint(0, self.num_negative - 1)
+            data_dict = self.negative_patches[patch_idx]
+
+        # Apply transform if provided
+        if self.transform:
+            # Apply transform to patch data if needed
+            for key, value in data_dict.items():
+                if hasattr(value, 'shape'):  # Apply to array-like data
+                    data_dict[key] = self.transform(value)
+
+        return data_dict
